fix: RESETHACK for CT5880
Some revisions of the Ensoniq 1371 labelled CT5880 require a certain bit to be set to get the AC97 out of hard reset. These revisions are correctly identified and the RESETHACK flag is set. The certain bit that needs to be set is located in the "Chip Select and Status Register" (CSSR), not in the "Serial Interface Control Register" (SCTRL).

This commit fixes SBEMU on a mainboard with an onboard Sound Blaster PCI.